qt6: Omit QWebkit and QtScript references when compiling for Qt6.
This commit essentially hacks out large pieces of code that are no
longer supported in Qt6.  It will cause anything that uses QWebKit or
QtScript to not function properly, if it functions at all.  This means
mythnetbrowser and the services API.  Removing all of this code does
allow the rest of MythTV to compile against the Qt6 toolkit.